admin_panel/requirements.txt
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-split-settings==1.1.0
django==4.0.4
flake8==4.0
isort==5.10.1
psycopg2-binary==2.9
pytest==7.1.2
python-dotenv==0.20.0
six==1.16.0
uWSGI==2.0.20
more-itertools==8.14.0
python-json-logger==2.0.4
django-log-request-id==2.1.0
sentry-sdk==1.14.0